updated to also cover click shortcuts

+++ code/modules/research/xenobiology/xenobio_camera.dm
@@ -437,8 +437,11 @@
 			var/mob/living/carbon/monkey/food = new /mob/living/carbon/monkey(T, TRUE, C)
 			if (!QDELETED(food))
 				food.LAssailant = C
-				X.monkeys --
-				to_chat(C, "[X] now has [X.monkeys] monkeys left.")
+				X.monkeys--
+				X.monkeys = round(X.monkeys, 0.1)		//Prevents rounding errors
+				to_chat(C, "[X] now has [X.monkeys] monkeys stored.")
+		else
+			to_chat(C, "[X] needs to have at least 1 monkey stored. Currently has [X.monkeys] monkeys stored.")
 
 //Pick up monkey
 /obj/machinery/computer/camera_advanced/xenobio/proc/XenoMonkeyClickCtrl(mob/living/user, mob/living/carbon/monkey/M)
@@ -449,10 +452,15 @@
 	var/mob/camera/aiEye/remote/xenobio/E = C.remote_control
 	var/obj/machinery/computer/camera_advanced/xenobio/X = E.origin
 	var/area/mobarea = get_area(M.loc)
+	if(!X.connected_recycler)
+		to_chat(C, "<span class='notice'>There is no connected monkey recycler.  Use a multitool to link one.</span>")
+		return
 	if(mobarea.name == E.allowed_area || mobarea.xenobiology_compatible)
 		if(!M.stat)
 			return
 		M.visible_message("[M] vanishes as [p_theyre()] reclaimed for recycling!")
-		X.monkeys = round(X.monkeys + 0.2, 0.1)
-		to_chat(C, "[X] now has [X.monkeys] monkeys available.")
+		X.connected_recycler.use_power(500)
+		X.monkeys += connected_recycler.cube_production
+		X.monkeys = round(X.monkeys, 0.1)		//Prevents rounding errors	
 		qdel(M)
+		to_chat(C, "[X] now has [X.monkeys] monkeys available.")
